+++ Gems/Atom/Feature/Common/Code/Source/ReflectionScreenSpace/ReflectionScreenSpaceBlurChildPass.cpp
@@ -38,7 +38,7 @@ namespace AZ
             if (m_imageSize != size)
             {
                 m_imageSize = size;
-                m_outputScale = (m_passType == PassType::Vertical) ? pow(2.0f, m_mipLevel) : 1.0f;
+                m_outputScale = (m_passType == PassType::Vertical) ? static_cast<float>(pow(2.0f, m_mipLevel)) : 1.0f;
 
                 m_updateSrg = true;
             }
